Modified to generate the desired RDF triples for dataset.
Use "id" (alias of "@id") to specify the ID of the RDF subject, then use "type" (alias of "@type") to specify the data type.
Don't alias "dataset" to {"@id": "https://id.gs1.org/dataset/{refID}"} if  https://id.gs1.org/dataset/{refID} is not intended as an RDF property or predicate.  It looks as though it's more likely to be the ID of an RDF Subject.
Added aliases for schema and dcat vocabularies.